backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
celery==5.4.0
redis==5.0.3
sqlalchemy[asyncio]==2.0.30
pgvector==0.2.5
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
requests==2.32.3
# Security
slowapi==0.1.9
structlog==24.4.0
# Migrations
alembic==1.18.4
# External APIs / AI
httpx==0.28.1
PyJWT==2.11.0
cryptography==46.0.5
openai==1.50.2
